+++ circuits/src/instructions/assertions.rs
@@ -125,7 +125,7 @@ pub(crate) mod tests {
         instructions::{AssertionInstructions, AssignmentInstructions},
         testing_utils::{FromScratch, Sampleable},
         types::InnerConstants,
-        utils::circuit_modeling::circuit_to_json,
+        utils::circuit_modeling::{circuit_to_json, cost_measure_end, cost_measure_start},
     };
 
     #[derive(Clone, Debug)]
@@ -183,6 +183,7 @@ pub(crate) mod tests {
             let x = chip.assign(&mut layouter, Value::known(self.x.clone()))?;
             let y = chip.assign_fixed(&mut layouter, self.y.clone())?;
 
+            cost_measure_start(&mut layouter);
             match self.operation {
                 Operation::Eq => chip.assert_equal(&mut layouter, &x, &y),
                 Operation::Neq => chip.assert_not_equal(&mut layouter, &x, &y),
@@ -191,6 +192,7 @@ pub(crate) mod tests {
                     chip.assert_not_equal_to_fixed(&mut layouter, &x, self.y.clone())
                 }
             }?;
+            cost_measure_end(&mut layouter);
 
             chip.load_from_scratch(&mut layouter)
         }