+++ circuits/src/instructions/canonicity.rs
@@ -137,7 +137,10 @@ pub(crate) mod tests {
     use crate::{
         instructions::{AssertionInstructions, AssignmentInstructions},
         types::InnerValue,
-        utils::{circuit_modeling::circuit_to_json, util::FromScratch},
+        utils::{
+            circuit_modeling::{circuit_to_json, cost_measure_end, cost_measure_start},
+            util::FromScratch,
+        },
     };
 
     #[derive(Clone, Debug)]
@@ -201,11 +204,13 @@ pub(crate) mod tests {
                 .collect::<Result<Vec<_>, Error>>()?;
             let bound = self.bound.clone();
 
+            cost_measure_start(&mut layouter);
             let res = match self.operation {
                 Operation::Canonical => chip.is_canonical(&mut layouter, &bits),
                 Operation::Lower => chip.le_bits_lower_than(&mut layouter, &bits, bound),
                 Operation::Geq => chip.le_bits_geq_than(&mut layouter, &bits, bound),
             }?;
+            cost_measure_end(&mut layouter);
 
             chip.assert_equal_to_fixed(&mut layouter, &res, self.expected)?;