+++ circuits/src/instructions/division.rs
@@ -159,7 +159,9 @@ pub(crate) mod tests {
 
     use super::*;
     use crate::{
-        testing_utils::FromScratch, types::InnerValue, utils::circuit_modeling::circuit_to_json,
+        testing_utils::FromScratch,
+        types::InnerValue,
+        utils::circuit_modeling::{circuit_to_json, cost_measure_end, cost_measure_start},
     };
 
     struct TestCircuit<F, Assigned, DivChip>
@@ -203,7 +205,9 @@ pub(crate) mod tests {
             let chip = DivChip::new_from_scratch(&config);
 
             let x = chip.assign(&mut layouter, self.dividend)?;
+            cost_measure_start(&mut layouter);
             let (q, r) = chip.div_rem(&mut layouter, &x, self.divisor.clone(), None)?;
+            cost_measure_end(&mut layouter);
 
             chip.assert_equal_to_fixed(&mut layouter, &q, self.expected.0)?;
             chip.assert_equal_to_fixed(&mut layouter, &r, self.expected.1)?;